Reenable float aggregation pushdown for duckdb except for min()

+++ vortex-bench/sql/vortex/init.sql
@@ -3,7 +3,7 @@
 COPY (
     SELECT
         i AS id,
-        i % 1000 AS col,
-        (i * 2654435761) % 100000 AS col2
+        (i % 1000)::INTEGER AS col,
+        ((i * 2654435761) % 100000)::INTEGER AS col2
     FROM range(2500000000) t(i)
 ) TO 'test.parquet' (FORMAT parquet);
